+++ src/libs/WixToolset.Data/Messaging.cs
@@ -11,9 +11,6 @@ namespace WixToolset.Data
 {
     using System;
     using System.Collections.Generic;
-    using System.Globalization;
-    using System.Linq;
-    using System.Text;
 
     public class Messaging : IMessageHandler
     {