+++ History.md
@@ -1,3 +1,5 @@
+* FireGiant: WIXBUG:4561 - update WixCop to migrate v3 source code to v4.0.
+
 * RobMen: WIXBUG:4565 - add WixToolset.Data.dll and WixToolset.Extensibility.dll to binaries.zip.
 
 ## WixBuild: Version 4.0.2220.0